Move gfs API to separate .cpp file
Fixes: #424
